Rename and restructure python example.
Arrange code into python packages, and rename modules/folders according to PEP 8.
