+++ examples/python/Dockerfile.columnoperations
@@ -6,4 +6,4 @@ WORKDIR /app
 RUN pip install --trusted-host pypi.python.org -r requirements.txt
 
 EXPOSE 50053
-CMD ["python", "ColumnOperations/ExtensionService_column.py"]
+CMD ["python", "ColumnOperations"]
