+++ examples/python/Dockerfile.helloworld
@@ -6,4 +6,4 @@ WORKDIR /app
 RUN pip install --trusted-host pypi.python.org -r requirements.txt
 
 EXPOSE 50052
-CMD ["python", "HelloWorld/ExtensionService_helloworld.py"]
+CMD ["python", "helloworld"]
