+++ examples/python/Dockerfile.scriptsupport
@@ -6,4 +6,4 @@ WORKDIR /app
 RUN pip install --trusted-host pypi.python.org -r requirements.txt
 
 EXPOSE 50051
-CMD ["python", "FullScriptSupport/ExtensionService_Script.py"]
+CMD ["python", "FullScriptSupport"]
