+++ examples/python/GetStarted.md
@@ -14,10 +14,10 @@ The following table includes a short description of each example and the functio
 | __Full script support using Pandas__ | Script | Numeric, String, Dual | Using the Pandas library and exec method to evaluate the script instead of eval. Also includes an example of writing a `TableDescription` in the script to be evaluated, when using the `Load ... Extension ...` statement. Otherwise it is the same example as the original full script support with all script functions supported. |
 
 For details about a particular example, see its documentation:
-- [Hello world](HelloWorld/README.md)
-- [Column operations](ColumnOperations/README.md)
-- [Full script support](FullScriptSupport/README.md)
-- [Full script support using Pandas](FullScriptSupport_Pandas/README.md)
+- [Hello world](helloworld/README.md)
+- [Column operations](columnoperations/README.md)
+- [Full script support](fullscriptsupport/README.md)
+- [Full script support using Pandas](fullscriptsupport_pandas/README.md)
 
 ## Running the Python examples
 Follow these steps to quickly set up and run an example of your choice on your local machine, with an insecure connection, using either Qlik Sense Desktop, Qlik Sense Enterprise, QlikView Desktop or QlikView Server. To run several examples, or to run the examples with a secure connection or on another node, configure your system according to the instructions referenced in [Configuring SSE plugins in Qlik](../../docs/configuration.md). The `<examplename>`, `<EngineName>` and `<port>` referred to below are mapped to each example as follows:
@@ -25,18 +25,18 @@ Follow these steps to quickly set up and run an example of your choice on your l
 | __Example__ | __`<examplename>`__ | __`<EngineName>`__ | __`<port>`__ |
 | ----- | ----- | ----- | ----- |
 | __Hello world__ | helloworld | HelloWorld | 50052 |
-| __Column operations__ | column | Column | 50053 |
-| __Full script support__ | script | Script | 50051 |
-| __Full script support using Pandas__ | scriptPandas | ScriptPandas | 50056 |
+| __Column operations__ | columnsupport | Column | 50053 |
+| __Full script support__ | fullscriptsupport | Script | 50051 |
+| __Full script support using Pandas__ | fullscriptsupport_pandas | ScriptPandas | 50056 |
 
 ### Qlik Sense Desktop
 1. Install Qlik Sense Desktop (June 2017 release or later).
 2. Make sure you have Python 3.4 (or later) installed as well as the `grpcio` package. For more information, see [Prerequisites for running the Python examples](prerequisites.md).
 3. Add `SSEPlugin=<EngineName>,localhost:<port>` on a new line in your *Settings.ini* file located at *C:\Users\\[user]\Qlik\Sense*. Insert the values for `<EngineName>` and `<port>` from the table above for the selected example.
 4. Copy the *.qvf* file from the selected example folder to *C:\Users\\[user]\Qlik\Sense\Apps*.
-5. Run the corresponding `ExtensionService_<examplename>.py` file. The easiest way to do this is to open a command prompt, go to the selected example folder and type:
+5. Run the corresponding `<examplename>` python package. The easiest way to do this is to open a command prompt, go to the example\python folder and type:
 
-   `python ExtensionService_<examplename>.py`  
+   `python <examplename>`  
 
    Insert the value for `<examplename>` from the table above for the selected example.
 6. Start Qlik Sense Desktop and open the app for the example you chose.
@@ -50,9 +50,9 @@ Follow these steps to quickly set up and run an example of your choice on your l
 
     __Note:__ Configuring the plugin via the *settings.ini* file is only valid for this node/machine; the settings will not be visible in the QMC. Also no access rules would apply in this case.
 4. Add the *.qvf* file from the selected example folder to QMC.
-5. Run the corresponding `ExtensionService_<examplename>.py` file. The easiest way to do this is to open a command prompt, go to folder for the selected example and type:
+5. Run the corresponding `<examplename>` python package. The easiest way to do this is to open a command prompt, go to the example\python folder and type:
 
-   `python ExtensionService_<examplename>.py`
+   `python <examplename>`
 
    Insert the value for `<examplename>` from the table above for the selected example.
 6. Start Qlik Sense Enterprise and open the app for the example you chose.
@@ -61,9 +61,9 @@ Follow these steps to quickly set up and run an example of your choice on your l
 1. Install QlikView Desktop (November 2017 release or later).
 2. Make sure you have Python 3.4 (or later) installed as well as the `grpcio` package. For more information, see [Prerequisites for running the Python examples](prerequisites.md).
 3. Add `SSEPlugin=<EngineName>,localhost:<port>` on a new line in your *Settings.ini* file, below the heading [Settings 7]. *Settings.ini* is located at *C:\Users\\[user]\AppData\Roaming\QlikTech\QlikView*. Insert the values for `<EngineName>` and `<port>` from the table above for the selected example.
-4. Run the corresponding `ExtensionService_<examplename>.py` file. The easiest way to do this is to open a command prompt, go to the selected example folder and type:
+4. Run the corresponding `<examplename>` python package. The easiest way to do this is to open a command prompt, go to the example\python folder and type:
 
-   `python ExtensionService_<examplename>.py`  
+   `python <examplename>`  
 
    Insert the value for `<examplename>` from the table above for the selected example.
 6. Start QlikView Desktop and open the *.qvw* file in the selected example folder.
@@ -73,9 +73,9 @@ Follow these steps to quickly set up and run an example of your choice on your l
 2. Make sure you have Python 3.4 (or later) installed as well as the `grpcio` package. For more information, see [Prerequisites for running the Python examples](prerequisites.md).
 3. Add `SSEPlugin=<EngineName>,localhost:<port>` on a new line in your *Settings.ini* file, below the heading [Settings 7]. *Settings.ini* is located at *C:\ProgramData\QlikTech\QlikViewServer*. Insert the values for `<EngineName>` and `<port>` from the table above for the selected example.
 4. Add the *.qvw* file from the selected example folder to your document root (e.g. *C:\ProgramData\QlikTech\Documents*) or a mounted folder.
-5. Run the corresponding `ExtensionService_<examplename>.py` file. The easiest way to do this is to open a command prompt, go to folder for the selected example and type:
+5. Run the corresponding `<examplename>` python package. The easiest way to do this is to open a command prompt, go to the example\python folder and type:
 
-   `python ExtensionService_<examplename>.py`
+   `python <examplename>`
 
    Insert the value for `<examplename>` from the table above for the selected example.
 6. Start QlikView Server and open the app for the example you chose.
