+++ examples/python/columnoperations/.gitignore
