+++ examples/python/columnoperations/SSE_Column_Operations.qvw
