+++ examples/python/columnoperations/logger.config
