+++ examples/python/columnoperations/ssedata.py
