+++ examples/python/fullscriptsupport/.gitignore
