+++ examples/python/fullscriptsupport/SSE_Full_Script_Support.qvw
