+++ examples/python/fullscriptsupport/__main__.py
@@ -9,11 +9,11 @@ from concurrent import futures
 
 # Add Generated folder to module path.
 PARENT_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-sys.path.append(os.path.join(PARENT_DIR, 'Generated'))
+sys.path.append(os.path.join(PARENT_DIR, 'generated'))
 
 import ServerSideExtension_pb2 as SSE
 import grpc
-from ScriptEval_script import ScriptEval
+from scripteval import ScriptEval
 
 _ONE_DAY_IN_SECONDS = 60 * 60 * 24
