+++ examples/python/fullscriptsupport/logger.config
