+++ examples/python/fullscriptsupport/ssedata.py
