+++ examples/python/fullscriptsupport_pandas/.gitignore
