+++ examples/python/fullscriptsupport_pandas/logger.config
