+++ examples/python/fullscriptsupport_pandas/ssedata.py
