+++ examples/python/helloworld/.gitignore
