+++ examples/python/helloworld/__main__.py
@@ -11,12 +11,12 @@ from datetime import datetime
 
 # Add Generated folder to module path.
 PARENT_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-sys.path.append(os.path.join(PARENT_DIR, 'Generated'))
+sys.path.append(os.path.join(PARENT_DIR, 'generated'))
 
 import ServerSideExtension_pb2 as SSE
 import grpc
-from SSEData_helloworld import FunctionType
-from ScriptEval_helloworld import ScriptEval
+from ssedata import FunctionType
+from scripteval import ScriptEval
 
 _ONE_DAY_IN_SECONDS = 60 * 60 * 24
 
@@ -305,11 +305,11 @@ if __name__ == '__main__':
     parser = argparse.ArgumentParser()
     parser.add_argument('--port', nargs='?', default='50052')
     parser.add_argument('--pem_dir', nargs='?')
-    parser.add_argument('--definition_file', nargs='?', default='FuncDefs_helloworld.json')
+    parser.add_argument('--definition_file', nargs='?', default='functions.json')
     args = parser.parse_args()
 
     # need to locate the file when script is called from outside it's location dir.
-    def_file = os.path.join(os.path.dirname(os.path.realpath(sys.argv[0])), args.definition_file)
+    def_file = os.path.join(os.path.dirname(os.path.abspath(__file__)), args.definition_file)
 
     calc = ExtensionService(def_file)
     calc.Serve(args.port, args.pem_dir)
