+++ examples/python/helloworld/ssedata.py
