+++ examples/python/requirements.txt
@@ -1,3 +1,4 @@
 grpcio==1.7.3
 numpy==1.13.1
 nose==1.3.7
+pandas==0.22.0
