+++ examples/python/test/test_columnoperations.py
@@ -6,7 +6,7 @@ import sys
 
 # Add Generated folder to module path.
 PARENT_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
-sys.path.append(os.path.join(PARENT_DIR, 'Generated'))
+sys.path.append(os.path.join(PARENT_DIR, 'generated'))
 
 import grpc
 import ServerSideExtension_pb2 as SSE
